requirements.txt
fastapi[standard]==0.115.6
pydantic-settings==2.7.0
langchain==0.3.25
langchain-community==0.3.25
langgraph==0.4.10
psycopg2-binary==2.9.10
pymysql==1.1.1
openai==1.86.0
redis==6.2.0
boto3==1.38.36
langdetect==1.0.9